refactor: presence direct-message 파일 구조 정리
